+++ ghost/core/core/server/api/endpoints/utils/serializers/input/utils/settings-filter-type-group-mapper.js
