+++ ghost/core/core/server/api/endpoints/utils/serializers/input/utils/settings-key-type-mapper.js
